% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aftprobiths.R
\name{aftprobiths}
\alias{aftprobiths}
\title{Function to implement the horseshoe shrinkage prior in integrated survival and binary regression}
\usage{
aftprobiths(ct, z, X, burn = 1000, nmc = 5000, thin = 1,
  alpha = 0.05, Xtest = NULL, cttest = NULL, ztest = NULL)
}
\arguments{
\item{ct}{survival response, a \eqn{n*2} matrix with first column as response and second column as right censored indicator,
1 is event time and 0 is right censored.}

\item{z}{binary response, a \eqn{n*1} vector with numeric values 0 or 1.}

\item{X}{Matrix of covariates, dimension \eqn{n*p}.}

\item{burn}{Number of burn-in MCMC samples. Default is 1000.}

\item{nmc}{Number of posterior draws to be saved. Default is 5000.}

\item{thin}{Thinning parameter of the chain. Default is 1 (no thinning).}

\item{alpha}{Level for the credible intervals. For example, alpha = 0.05 results in
95\% credible intervals.}

\item{Xtest}{test design matrix.}

\item{cttest}{test survival response.}

\item{ztest}{test binary response.}
}
\description{
Function to implement the horseshoe shrinkage prior in integrated survival and binary regression
}
\examples{
\dontrun{ #
}

}
\references{
Maity, A. K., Carroll, R. J., and Mallick, B. K. (2018),
Integration of Survival and Binary Data for Variable Selection and Prediction: A Bayesian Approach.
}
