% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aftprobiths.R
\name{aftprobiths}
\alias{aftprobiths}
\title{Function to implement the horseshoe shrinkage prior in integrated survival and binary regression}
\usage{
aftprobiths(ct, z, X, burn = 1000, nmc = 5000, thin = 1,
  alpha = 0.05, Xtest = NULL, cttest = NULL, ztest = NULL)
}
\arguments{
\item{ct}{survival response, a \eqn{n*2} matrix with first column as response and second column as right censored indicator,
1 is event time and 0 is right censored.}

\item{z}{binary response, a \eqn{n*1} vector with numeric values 0 or 1.}

\item{X}{Matrix of covariates, dimension \eqn{n*p}.}

\item{burn}{Number of burn-in MCMC samples. Default is 1000.}

\item{nmc}{Number of posterior draws to be saved. Default is 5000.}

\item{thin}{Thinning parameter of the chain. Default is 1 (no thinning).}

\item{alpha}{Level for the credible intervals. For example, alpha = 0.05 results in
95\% credible intervals.}

\item{Xtest}{test design matrix.}

\item{cttest}{test survival response.}

\item{ztest}{test binary response.}
}
\value{
\item{Beta.sHat}{Posterior mean of \eqn{\beta} for survival model, a \eqn{p} by 1 vector.}
\item{Beta.bHat}{Posterior mean of \eqn{\beta} for binary model, a \eqn{p} by 1 vector.}
\item{LeftCI.s}{The left bounds of the credible intervals for Beta.sHat.}
\item{RightCI.s}{The right bounds of the credible intervals for Beta.sHat.}
\item{LeftCI.b}{The left bounds of the credible intervals for Beta.bHat.}
\item{RightCI.b}{The right bounds of the credible intervals for Beta.bHat.}
\item{Beta.sMedian}{Posterior median of \eqn{beta} for survival model, a \eqn{p} by 1 vector.}
\item{Beta.bMedian}{Posterior median of \eqn{beta} for binary model, a \eqn{p} by 1 vector.}
\item{SigmaHat}{Posterior mean of variance covariance matrix.}
\item{LambdaHat}{Posterior mean of \eqn{\lambda}, a \eqn{p*1} vector.}
\item{TauHat} = {Posterior mean of \eqn{\tau}, a \eqn{2*1} vector.}
\item{Beta.sSamples}{Posterior samples of \eqn{\beta} for survival model.}
\item{Beta.bSamples}{Posterior samples of \eqn{\beta} for binary model.}
\item{LambdaSamples}{Posterior samples of \eqn{\lambda}.}
\item{TauSamples}{Posterior samples of \eqn{\tau}.}
\item{SigmaSamples}{Posterior samples of variance covariance matrix.}
\item{DIC.s}{DIC for survival model.}
\item{DIC.b}{DIC for binary model.}
\item{SurvivalHat}{Predictive survival probability.}
\item{LogTimeHat}{Predictive log time.}
}
\description{
Function to implement the horseshoe shrinkage prior in integrated survival and binary regression
}
\examples{
\dontrun{ #
}

}
\references{
Maity, A. K., Carroll, R. J., and Mallick, B. K. (2018),
Integration of Survival and Binary Data for Variable Selection and Prediction: A Bayesian Approach.
}
